refactor(ts): Move payloadFingerprint out of core library into example
Mirrors the Go-side change — payloadFingerprint is an application-level
convenience for 409 Conflict detection, not a protocol primitive.
