+++ typescript/packages/extensions/src/payment-identifier/utils.ts
@@ -2,8 +2,6 @@
  * Utility functions for the Payment-Identifier Extension
  */
 
-import { createHash } from "crypto";
-import type { PaymentPayload } from "@x402/core/types";
 import { PAYMENT_ID_MIN_LENGTH, PAYMENT_ID_MAX_LENGTH, PAYMENT_ID_PATTERN } from "./types";
 
 /**
@@ -28,29 +26,6 @@ export function generatePaymentId(prefix: string = "pay_"): string {
   // Generate UUID v4 without hyphens (32 hex chars)
   const uuid = crypto.randomUUID().replace(/-/g, "");
   return `${prefix}${uuid}`;
-}
-
-/**
- * Computes a deterministic fingerprint of a PaymentPayload.
- * This allows detecting whether two payloads with the same payment ID carry
- * identical or different content (for 409 Conflict detection).
- *
- * @param payload - The payment payload to fingerprint
- * @returns Hex-encoded SHA-256 hash of the canonical payload
- */
-export function payloadFingerprint(payload: PaymentPayload): string {
-  // JSON.stringify with a replacer that sorts keys at every level for determinism
-  const canonical = JSON.stringify(payload, (_key, value) => {
-    if (value !== null && typeof value === "object" && !Array.isArray(value)) {
-      const sorted: Record<string, unknown> = {};
-      for (const k of Object.keys(value as Record<string, unknown>).sort()) {
-        sorted[k] = (value as Record<string, unknown>)[k];
-      }
-      return sorted;
-    }
-    return value;
-  });
-  return createHash("sha256").update(canonical).digest("hex");
 }
 
 /**
